refactor(auth): move signup/login logic to auth module and clean up debug logs
