+++ .env.example
@@ -3,3 +3,5 @@ DB_HOST=
 DB_USER=
 DB_PASSWORD=
 DB_NAME=
+JWT_SECRET=
+REFRESH_TOKEN_SECRET=
