+++ config/database.js
@@ -1,10 +1,5 @@
 const { Sequelize } = require('sequelize');
 
-console.log('DB_NAME:', process.env.DB_NAME);
-console.log('DB_USER:', process.env.DB_USER);
-console.log('DB_PASSWORD:', process.env.DB_PASSWORD);
-console.log('DB_HOST:', process.env.DB_HOST);
-
 const sequelize = new Sequelize(
   process.env.DB_NAME,
   process.env.DB_USER,
